[windows] Add windows software DSLs
This (big big big) commit updates most of our software definitions to
make them Windows compatible. In most cases, an `if` switch has been
added to use a different path, a different installation source.

The main difference is that we do not compile on Windows. Not that it
wouldn't be possible in the long term but Omnibus doesn't make it
simpler than launching shell commands for every compilation step. In
addition to that, different osftware requirements support different
compilers (Python is officialy supported only if compiled with Visual
Studio, Curl has to be compiled with MinGW...).

So what we do instead, for most cases, is just download an MSI installer
and install it in the appropriate location from the command line with
`msiexec`. For `pip` packages that don't compile on Windows (for those
who need to compile something) we use wheels from Christoph Gohlke's
awesome website: http://www.lfd.uci.edu/~gohlke/pythonlibs/ .

Finally, for `pip` packages that compile on Windows or don't require any
compilation steps, we just use the embedded `pip` to retrieve them.

+++ config/software/curl.rb
@@ -18,37 +18,52 @@
 name "curl"
 default_version "7.41.0"
 
-dependency "zlib"
-dependency "openssl"
+if ohai['platform'] != 'windows'
+  source :url => "http://curl.haxx.se/download/curl-#{version}.tar.gz",
+         :md5 => '7321a0a3012f8eede729b5a44ebef5bd'
 
-source :url => "http://curl.haxx.se/download/curl-#{version}.tar.gz",
-       :md5 => "7321a0a3012f8eede729b5a44ebef5bd"
+  dependency "zlib"
+  dependency "openssl"
 
-relative_path "curl-#{version}"
+  relative_path "curl-#{version}"
 
-build do
-  ship_license "https://raw.githubusercontent.com/bagder/curl/master/COPYING"
-  block do
-    FileUtils.rm_rf(File.join(project_dir, 'src/tool_hugehelp.c'))
+  build do
+    ship_license "https://raw.githubusercontent.com/bagder/curl/master/COPYING"
+    block do
+      FileUtils.rm_rf(File.join(project_dir, 'src/tool_hugehelp.c'))
+    end
+
+    command ["./configure",
+             "--prefix=#{install_dir}/embedded",
+             "--disable-manual",
+             "--disable-debug",
+             "--enable-optimize",
+             "--disable-ldap",
+             "--disable-ldaps",
+             "--disable-rtsp",
+             "--enable-proxy",
+             "--disable-dependency-tracking",
+             "--enable-ipv6",
+             "--without-libidn",
+             "--without-gnutls",
+             "--without-librtmp",
+             "--with-ssl=#{install_dir}/embedded",
+             "--with-zlib=#{install_dir}/embedded"].join(" ")
+
+    command "make -j #{workers}", :env => {"LD_RUN_PATH" => "#{install_dir}/embedded/lib"}
+    command "make install"
   end
 
-  command ["./configure",
-           "--prefix=#{install_dir}/embedded",
-           "--disable-manual",
-           "--disable-debug",
-           "--enable-optimize",
-           "--disable-ldap",
-           "--disable-ldaps",
-           "--disable-rtsp",
-           "--enable-proxy",
-           "--disable-dependency-tracking",
-           "--enable-ipv6",
-           "--without-libidn",
-           "--without-gnutls",
-           "--without-librtmp",
-           "--with-ssl=#{install_dir}/embedded",
-           "--with-zlib=#{install_dir}/embedded"].join(" ")
+else
+  # Compiling is hard... let's ship binaries instead : TODO: react according to platform
+  source :url => "https://s3.amazonaws.com/dd-agent-omnibus/curl4-7.43.0.tar.gz",
+         :md5 => '885daa917d96c9d8278bda39a9295f47'
 
-  command "make -j #{workers}", :env => {"LD_RUN_PATH" => "#{install_dir}/embedded/lib"}
-  command "make install"
+  relative_path "curl"
+
+  build do
+    ship_license "https://raw.githubusercontent.com/bagder/curl/master/COPYING"
+
+    copy 'cygcurl-4.dll', "\"#{windows_safe_path(install_dir)}\\embedded\\Lib\\cygcurl.dll\""
+  end
 end
